Begin laying out TakedownRequest components
