+++ web/components/CreateTakedownRequestForm.tsx
@@ -1,0 +1,11 @@
+import { Entity } from "types";
+
+interface Props {
+	entity: Entity;
+}
+
+const CreateTakedownRequestForm = ({ entity }: Props) => {
+	return <div>Form goes here</div>;
+};
+
+export default CreateTakedownRequestForm;
