+++ web/pages/entities/places/[slug].tsx
@@ -11,8 +11,8 @@ import Layout from "components/Layout";
 import LoadingBlock from "components/LoadingBlock";
 import AudioItemComponent from "components/AudioItem";
 import TagWithRelationshipToObject from "components/TagWithRelationshipToObject";
-import AddTag from "components/AddTag";
-import EditTags from "components/EditTags";
+import AddTagButton from "components/AddTagButton";
+import EditTagsButton from "components/EditTagsButton";
 
 const PLACE_QUERY = gql`
 	query Place($slug: String!) {
@@ -136,11 +136,11 @@ const ViewPlaceBySlug = () => {
 						/>
 					))}
 					<div>
-						<AddTag entity={place} />
+						<AddTagButton entity={place} />
 						{tags.length > 0 && (
 							<>
 								<span className="text-gray-500 px-2">/</span>
-								<EditTags entity={place} />
+								<EditTagsButton entity={place} />
 							</>
 						)}
 					</div>
